// src/app/generated/dcc/model/langTextPair.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LangTextPair { 
    id?: string;
    refIds?: Array<string>;
    refTypes?: Array<string>;
    lang?: string;
    text?: string;
}